refactor(ui): split update status builders
Closes #2493
